object-detection/requirements.txt
flask
google-cloud-videointelligence
firebase-admin
